back-end/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pgvector
python-dotenv
psycopg2-binary
pydantic-settings
openai
tavily-python
mcp
langgraph
langchain-openai
langchain-core
greenlet
